Move hompage methods to seperate page. Organize imports
