+++ shuttle/driver_check_in/__init__.py
@@ -1,5 +1,6 @@
 from flask import render_template, session
 from flask_classy import FlaskView, route, request
+
 from shuttle.db import db_functions as db
 from shuttle.schedules.google_sheets_controller import SheetsController
 from shuttle.shuttle_controller import ShuttleController
