+++ shuttle/driver_logs/__init__.py
@@ -1,6 +1,8 @@
+from datetime import datetime
+
 from flask import render_template
 from flask_classy import FlaskView, route, request
-from datetime import datetime
+
 from shuttle.schedules.shuttle_schedules_controller import ScheduleController
 from shuttle.shuttle_controller import ShuttleController
